moves PoseidonParameters to utils/type.rs

+++ src/utils/bls381_x17_3.rs
@@ -13,9 +13,9 @@
 // 0x73eda753299d7d483339d80809a1d80553bda402fffe5bfeffffffff00000001
 use crate::poseidon::sbox::PoseidonSbox;
 
-pub const FULL_ROUNDS: usize = 8;
-pub const PARTIAL_ROUNDS: usize = 57;
-pub const WIDTH: usize = 3;
+pub const FULL_ROUNDS: u8 = 8;
+pub const PARTIAL_ROUNDS: u8 = 57;
+pub const WIDTH: u8 = 3;
 pub const SBOX: PoseidonSbox = PoseidonSbox::Exponentiation(17);
 
 pub const ROUND_CONSTS: [&str; 123] = [
